Deleted everything about shelves to remove a layer of the app for more simplicity. Also changed up some styling.
