+++ MyVirtualBookshelf/BookshelfPage.xaml.cs
@@ -1,107 +1,214 @@
 using MyVirtualBookshelf.Models;
 using System.Collections.ObjectModel;
-using System.Globalization;
-namespace MyVirtualBookshelf;
 #nullable disable
+namespace MyVirtualBookshelf;
 
 public partial class BookshelfPage : ContentPage
 {
+    
     private DatabaseHandler _db;
 
     /// <summary>
-    /// The shelves in the current bookshelf.
+    /// Google books service
     /// </summary>
-    public ObservableCollection<Shelf> Shelves { get; set; }
+    private BookService _bookService;
+    
+    /// <summary>
+    /// The books in this shelf.
+    /// </summary>
+    public ObservableCollection<Book> Books { get; set; }
 
     /// <summary>
-    /// The Id of the current bookshelf.
+    /// Used when EditBookBtn_Clicked is called
     /// </summary>
-    int BookshelfId { get; set; }
-    
-    public BookshelfPage(int bookshelfId)
-    {
-        this.BookshelfId = bookshelfId;
+    public Book BookToEdit { get; set; } = new Book();
+
+    /// <summary>
+    /// Used when DeleteBookBtn_Clicked is called
+    /// </summary>
+    public Book BookToDelete { get; set; } = new Book();
+
+    /// <summary>
+    /// Needed for when adding and deleting books from this shelf.
+    /// </summary>
+    public int ShelfId { get; set; }
+
+    /// <summary>
+    /// Reference to the bookshelf that this shelf resides in.
+    /// Required for updating BookCount property.
+    /// </summary>
+    public MainPage MainPageToUpdate { get; set; }
+
+    public BookshelfPage(int sid, MainPage bookshelfPage)
+	{
         InitializeComponent();
+        this.ShelfId = sid;
+        MainPageToUpdate = bookshelfPage;
 
-        // Display the shelves in the bookshelf 
+        // Google books service
+        _bookService = new BookService();
+
+        // Displays list of books on the page 
         _db = new DatabaseHandler();
-        Shelves = new ObservableCollection<Shelf>();
+        Books = new ObservableCollection<Book>();
 
-        PopulateBookshelf();
+        PopulateShelf();
         this.BindingContext = this;
     }
 
     /// <summary>
-    /// Updates the shelves display.
+    /// Calls the database and retrieves every book for the shelf via
+    /// the ShelfId property.
     /// </summary>
-    public void PopulateBookshelf()
+    public void PopulateShelf()
     {
-        Shelves.Clear();
+        Books.Clear();
 
-        List<Shelf> updatedShelves = _db.GetBookshelfContents(BookshelfId);
-        
-        int i = 1;
-        foreach (Shelf shelf in updatedShelves)
+        List<Book> updatedShelf = _db.GetBookshelfContents(ShelfId);
+
+        foreach (Book book in updatedShelf)
         {
-            // Create shelf name
-            shelf.ShelfName = "Shelf " + i.ToString();
-
-            DisplayShelfHint(shelf);
-
-            Shelves.Add(shelf);
-            i++;
+            Books.Add(book);
         }
     }
 
     /// <summary>
-    /// This method grabs all the books in a shelf, if any, and
-    /// creates a string that is used to display a hint at what
-    /// is in the shelf.
-    /// </summary>
-    /// <param name="shelf"></param>
-    public void DisplayShelfHint(Shelf shelf)
-    {       
-        List<Book> allBooks = _db.GetShelfContents(shelf.Id);
-
-        // Avoid null-reference exception by checking ahead
-        if (allBooks.Count == 0)
-        {
-            shelf.ShelfContentsHint = "This shelf is empty";
-        }
-        else
-        {
-            // Create a string using fencepost method           
-            for (int i = 0; i < allBooks.Count && i < 3; i++)
-            {
-                // Add the title to ShelfContentsHint
-                if (i == 0)
-                {
-                    shelf.ShelfContentsHint += allBooks[i].Title;
-                }
-                else
-                {
-                    shelf.ShelfContentsHint += ", " + allBooks[i].Title;
-                }
-            }
-        }       
-    }
-
-    /// <summary>
-    /// Opens the specified shelf.
+    /// When the search bar button is clicked, it adds the value
+    /// in the search bar to the current shelf.
     /// </summary>
     /// <param name="sender"></param>
     /// <param name="e"></param>
-    public async void OpenShelfBtn_Clicked(object sender, EventArgs e)
+    public async void BookSearchBarButton_Clicked(object sender, EventArgs e)
     {
-        // Ensure the sender is a Button
-        Button clickedButton = sender as Button;
-
-        // Check if the clickedButton is not null and safely access its BindingContext
-        if (clickedButton != null)
+        if (!string.IsNullOrEmpty(BookSearchbar.Text))
         {
-            Shelf selectedShelf = clickedButton.BindingContext as Shelf;
-            ShelfPage shelfPage = new ShelfPage(selectedShelf.Id, this);
-            await Navigation.PushAsync(shelfPage);
+            try
+            {
+                // throw new Exception(); // For debugging
+                string searchQuery = BookSearchbar.Text;
+                List<Book> searchResults = await _bookService.SearchGoogleBooksAsync(searchQuery);
+
+                // Add the first result from the Google Books query to the db and perform
+                // necessary updates.
+                if (searchResults.Count > 0)
+                {
+                    Book bookToAdd = searchResults[0];
+                    _db.AddBook(ShelfId, bookToAdd);
+                    BookSearchbar.Text = null;
+                    PopulateShelf();
+                    MainPageToUpdate.PopulateBookshelves();
+                }
+            }
+            catch (Exception ex)
+            {   // Api call failed - Add SearchBar text as a book
+                Book titleOnlyBook = new Book(ShelfId, BookSearchbar.Text);
+                _db.AddBook(ShelfId, titleOnlyBook);
+                BookSearchbar.Text = null;
+                PopulateShelf();
+                MainPageToUpdate.PopulateBookshelves();
+            }
         }
+    }
+
+    /// <summary>
+    /// Opens the delete menu and stores the specified book's title and id.
+    /// </summary>
+    /// <param name="sender"></param>
+    /// <param name="e"></param>
+    public void DeleteBookBtn_Clicked(object sender, EventArgs e)
+    {
+        if (!ConfirmMenu.IsVisible)
+        {
+            // Cast sender as a button to access BindingContext
+            Button clickedButton = sender as Button;
+
+            // Display BookTitle to the confirmation menu 
+            Book book = clickedButton.BindingContext as Book;
+            BookToDelete.Id = book.Id;
+            BookToDelete.Title = book.Title;
+
+            ConfirmDeleteLabel.Text = $"Delete \"{BookToDelete.Title}\" ?";
+
+            OpaqueBackground.IsVisible = true;
+            OpaqueBackground.Opacity = 0.3;
+            ConfirmMenu.IsVisible = true;
+        }
+    }
+
+    /// <summary>
+    /// Deletes the specified book and hides the delete menu.
+    /// </summary>
+    /// <param name="sender"></param>
+    /// <param name="e"></param>
+    public void ConfirmDeleteBtn_Clicked(object sender, EventArgs e)
+    {
+        ConfirmMenu.IsVisible = false;
+        OpaqueBackground.IsVisible = false;
+
+        _db.DeleteBook(ShelfId, BookToDelete.Id);
+        PopulateShelf();
+        MainPageToUpdate.PopulateBookshelves();
+    }
+
+    /// <summary>
+    /// Hides the delete menu.
+    /// </summary>
+    /// <param name="sender"></param>
+    /// <param name="e"></param>
+    public void DeclineDeleteBtn_Clicked(object sender, EventArgs e)
+    {
+        ConfirmMenu.IsVisible = false;
+        OpaqueBackground.IsVisible = false;
+    }
+
+    public void EditBookBtn_Clicked(object sender, EventArgs e)
+    {
+        if (!ConfirmMenu.IsVisible || !EditMenu.IsVisible)
+        {
+            // Cast sender as a button to access BindingContext
+            Button clickedButton = sender as Button;
+
+            // Display BookTitle to the confirmation menu 
+            Book book = clickedButton.BindingContext as Book;
+            BookToEdit.Id = book.Id;
+            BookToEdit.Title = book.Title;
+            BookToEdit.Author = book.Author;
+            BookToEdit.Genre = book.Genre;
+            BookToEdit.Isbn = book.Isbn;
+            BookToEdit.BookshelfId = book.BookshelfId;
+
+            OpaqueBackground.IsVisible = true;
+            OpaqueBackground.Opacity = 0.3;
+            EditMenu.IsVisible = true;
+
+            BookTitleEntry.Text = BookToEdit.Title;
+            BookAuthorEntry.Text = BookToEdit.Author;
+            BookGenreEntry.Text = BookToEdit.Genre;
+            BookIsbnEntry.Text = BookToEdit.Isbn;
+        }
+    }
+
+    public void SaveEditBtn_Clicked(object sender, EventArgs e)
+    {
+        Book updatedBook = new Book();
+        updatedBook.Id = BookToEdit.Id;
+        updatedBook.Title = BookTitleEntry.Text;
+        updatedBook.Author = BookAuthorEntry.Text;
+        updatedBook.Genre = BookGenreEntry.Text;
+        updatedBook.Isbn = BookIsbnEntry.Text;
+        updatedBook.BookshelfId = BookToEdit.BookshelfId;
+
+        _db.UpdateBook(updatedBook);
+
+        PopulateShelf();
+
+        EditMenu.IsVisible = false;
+        OpaqueBackground.IsVisible = false;
+    }
+
+    public void CancelEditBtn_Clicked(object sender, EventArgs e)
+    {
+        EditMenu.IsVisible = false;
+        OpaqueBackground.IsVisible = false;
     }
 }
