+++ MyVirtualBookshelf/Resources/Styles/Colors.xaml
@@ -6,12 +6,12 @@
 
     <!-- Note: For Android please see also Platforms\Android\Resources\values\colors.xml -->
 
-    <Color x:Key="Primary">#512BD4</Color>
-    <Color x:Key="PrimaryDark">#ac99ea</Color>
-    <Color x:Key="PrimaryDarkText">#242424</Color>
-    <Color x:Key="Secondary">#DFD8F7</Color>
-    <Color x:Key="SecondaryDarkText">#9880e5</Color>
-    <Color x:Key="Tertiary">#2B0B98</Color>
+    <Color x:Key="Primary">#FF7043</Color>
+    <Color x:Key="PrimaryDark">#D84315</Color>
+    <Color x:Key="PrimaryDarkText">#FFFFFF</Color>
+    <Color x:Key="Secondary">#FFAB91</Color>
+    <Color x:Key="SecondaryDarkText">#BF360C</Color>
+    <Color x:Key="Tertiary">#FF5722</Color>
 
     <Color x:Key="White">White</Color>
     <Color x:Key="Black">Black</Color>
